♻️ Factor out EtaCalculator
See #964
